UI Update
Added some more functionality to the UI. Created a play button, and the game won't start until its clicked. Added a quit button that instantly closes the game. Added a button that opens an options menu, but there are no options to modify atm, and the game doesn't pause when its open.