Life Points
Added life points and their corosponding UI to the game. The life points don't do anything when they hit zero, but they're visually indicated and tracked in code.
Also changed the core gameplay in accordance to the new gameplay loop plan. Score is not reset on "death", only a life is lost. Streaking just adds directly to the current score.